natlas-agent/requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
urllib3==1.24.1
